[NCL-4809] Add support for handling repositories-backup.xml
Code handling the backup file is now shared across PME and GME.
